set up a calibration script

+++ src/influenza_USA/SVIR/TDPF.py
@@ -14,7 +14,7 @@ def exponential_waning_function(t, states, param, waning_start, T_s):
     if t < waning_start:
         return 1.0
     else:
-        return np.exp(-1/T_s * (t - waning_start)/timedelta(days=1))
+        return np.exp(-1/T_s * float((t - waning_start)/timedelta(days=1)))
 
 class make_vaccination_function():
